usecase_code/10-azure-function/11_func_read_from_storage/requirements.txt
azure-functions
azure-identity>=1.15.0
azure-storage-blob>=12.19.0
